Changed API processes to reset index after dropping intra-regional migrants. Modified MK01 net migration visualisation.

+++ api/stat_fi_api_helpers.py
@@ -59,7 +59,7 @@ def create_dataframe_from_stat_fi_format(json_data):
 
 def fix_dataframe_dtypes(df):
     return df.astype({'Year': 'int64',
-                      'Value': 'float64'})
+                      'Value': 'int64'})
 
 def write_to_json_file(data, file_name):
     file_path = os.path.join('raw_data', file_name)
